3-req finish

+++ DatabaseInterfacing/Domain/DTOs/ReadingDto.cs
@@ -5,14 +5,14 @@ namespace DatabaseInterfacing.Domain.DTOs;
 public class ReadingDto
 {
     [JsonPropertyName("water_conductivity")]
-    public int? WaterConductivity { get; set; }
+    public float? WaterConductivity { get; set; }
     
     [JsonPropertyName("water_temperature")]
     public float? WaterTemperature { get; set; }
 
     
     [JsonConstructor]
-    public ReadingDto(int? waterConductivity, float? waterTemperature)
+    public ReadingDto(float? waterConductivity, float? waterTemperature)
     {
         WaterConductivity = waterConductivity;
         WaterTemperature = waterTemperature;
